Fix: binary search in ini_parse
If the size of the array is 2, it could fail because of getting to
position -1 which in unsigned is incorrect.

+++ src/ini_parser.c
@@ -40,11 +40,11 @@ static const struct cmd *find_cmd(const char *name) {
 }
 
 static const struct cmd_option *find_cmd_option(const struct cmd_opts *cmd_opts, const char *name) {
-	unsigned bottom = 0;
-	unsigned top = cmd_opts->size - 1;
+	int bottom = 0;
+	int top = (int)cmd_opts->size - 1;
 
 	while (bottom <= top) {
-		const unsigned mid = (bottom + top) / 2;
+		const int mid = (bottom + top) / 2;
 		const int cmp_res = strcmp(cmd_opts->names[mid], name);
 		if (cmp_res == 0)
 			return cmd_opts->opts + mid;
